[dagit] Virtualized Asset Catalog
