+++ js_modules/dagit/packages/core/src/assets/AssetTableFragment.tsx
@@ -1,0 +1,36 @@
+import {gql} from '@apollo/client';
+
+export const ASSET_TABLE_DEFINITION_FRAGMENT = gql`
+  fragment AssetTableDefinitionFragment on AssetNode {
+    id
+    groupName
+    isSource
+    partitionDefinition {
+      description
+    }
+    description
+    repository {
+      id
+      name
+      location {
+        id
+        name
+      }
+    }
+  }
+`;
+
+export const ASSET_TABLE_FRAGMENT = gql`
+  fragment AssetTableFragment on Asset {
+    __typename
+    id
+    key {
+      path
+    }
+    definition {
+      id
+      ...AssetTableDefinitionFragment
+    }
+  }
+  ${ASSET_TABLE_DEFINITION_FRAGMENT}
+`;
