+++ js_modules/dagit/packages/core/src/workspace/VirtualizedWorkspaceTable.tsx
@@ -43,10 +43,15 @@ export const RepoRow: React.FC<{
   );
 };
 
-export const LoadingOrNone: React.FC<{queryResult: QueryResult<any, any>}> = ({queryResult}) => {
+export const LoadingOrNone: React.FC<{
+  queryResult: QueryResult<any, any>;
+  noneString?: React.ReactNode;
+}> = ({queryResult, noneString = 'None'}) => {
   const {called, loading, data} = queryResult;
   return (
-    <div style={{color: Colors.Gray500}}>{!called || (loading && !data) ? 'Loading' : 'None'}</div>
+    <div style={{color: Colors.Gray500}}>
+      {!called || (loading && !data) ? 'Loading' : noneString}
+    </div>
   );
 };
